@@ -10,14 +10,14 @@
         }
     ],
     "require": {
-        "php": ">=7.1.0",
-        "divineomega/php-countries": "^2.1.1",
-        "psr/cache": "^1.0",
-        "guzzlehttp/guzzle": "^6.3|^7.0",
+        "php": "^8.1",
+        "divineomega/php-countries": "^3.0",
+        "psr/cache": "^1.0|^2.0|^3.0",
+        "guzzlehttp/guzzle": "^7.0",
         "ext-json": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0 || ^8.0",
+        "phpunit/phpunit": "^9.5",
         "php-coveralls/php-coveralls": "^2.1"
     },
     "autoload": {
