@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace DivineOmega\Geolocation\Interfaces;
 
@@ -8,7 +9,6 @@ interface LocationProviderInterface
 {
     /**
      * @param string $ip
-     * @return Country
      */
-    public function getCountryByIP(string $ip);
+    public function getCountryByIP(string $ip): ?Country;
 }
