@@ -1,8 +1,10 @@
 <?php
+declare(strict_types=1);
 
 namespace DivineOmega\Geolocation\LocationProviders;
 
 use DivineOmega\Geolocation\Interfaces\LocationProviderInterface;
+use DivineOmega\Countries\Country;
 use Exception;
 
 class FreeGeoIP implements LocationProviderInterface
@@ -17,8 +19,8 @@ class FreeGeoIP implements LocationProviderInterface
         throw new Exception('FreeGeoIP has shutdown, please use the IpStack location provider.');
     }
 
-    public function getCountryByIP(string $ip)
+    public function getCountryByIP(string $ip): ?Country
     {
-        //
+        return null;
     }
 }
